Added 2fa button to profile and made changes to signup
