+++ Frontend/src/index.html
@@ -49,6 +49,31 @@
         </div>
     </div>
     
+    <!-- Verification Code Modal 
+    <div class="modal fade" id="verificationModal" tabindex="-1" aria-labelledby="verificationLabel" aria-hidden="true">
+        <div class="modal-dialog">
+            <div class="modal-content">
+                <div class="modal-header">
+                    <h5 class="modal-title" id="verificationLabel">Enter Verification Code</h5>
+                    <button type="button" class="close" data-bs-dismiss="modal" aria-label="Close"></button>
+                </div>
+                <div class="modal-body">
+                    <p class="labelFont">A verification code has been sent to your email.</p>
+                    <form id="verificationForm">
+                        <div class="mb-3">
+                            <label for="verificationCode" class="labelFont">Verification Code</label>
+                            <input type="text" class="form-control" id="verificationCode" placeholder="Enter code" required>
+                        </div>
+                        <div class="d-flex justify-content-end">
+                            <button type="submit" class="submit btn btn-primary me-2">Verify</button>
+                            <button type="button" class="submit btn btn-secondary" data-bs-dismiss="modal">Cancel</button>
+                        </div>
+                    </form>
+                </div>
+            </div>
+        </div>
+    </div> -->
+    
 
     <!-- Pong Game Modal -->
     <div class="modal fade game-modal" id="pongModal" tabindex="-1" role="dialog" aria-labelledby="pongModalLabel" aria-hidden="true">
