+++ Frontend/src/scss/_modal.scss
@@ -57,9 +57,10 @@
 }
 
 .modal-body {
-  display: block;
+  display: flex;
   justify-content: center;
   align-items: center;
+  flex-direction: column;
 }
 
 .modal-title {
